Implement bit_qtree.{c,h}pp
Initial attempt at storing a quadtree for the purposes of a baseline to
compare against. However, is innapropriate, as is not itself queryable.
Will need to write another class to handle this.

On the upside, bit_qtree is a nice way to *store* quadtrees, and build
other representations of quadtrees from. So it will probably feature
heavily herein in the strategy of constructing other forms of trees.

+++ src/leeyang.hpp
@@ -67,7 +67,7 @@ n_qnode* new_qnode(void*);
 void free_qnode(n_qnode*, int);
 n_qtree* new_qtree(unsigned int);
 void free_qtree(n_qtree*, int);
-link_node* new_link_node(void*, unsigned int, unsigned int);
+link_node* new_link_node(void*, long unsigned int, long unsigned int);
 void free_link_node(link_node*);
 void insert_coord(n_qtree*, void*, unsigned int, unsigned int, int);
 void* query_coord(n_qtree*, unsigned int, unsigned int);
